# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: TeamLaunch Sync API
  version: 1.0.0
  description: |
    团队统一入口工具的局域网同步服务 API。服务内嵌在管理员客户端进程内（ADR-004）。

    安全：明文 HTTP 运行在局域网（无 PKI）。写操作必须由 `Bearer` 令牌 + HMAC 请求签名
    双重保护（ADR-006），该机制已由用户于 2026-09-29 确认，作为 PRD §14「TLS + JWT 或
    经架构师确认的等价机制」条款的正式替代；TLS（自签证书 + TOFU）保留为 P1 升级路径。

    信封：所有 JSON 响应统一 `{code, data, message}`，`code = 0` 表示成功。
    失败时 `code` 为非 0 整数，并附带 `error`（ERR_* 机器码）与 `diagnosticId`。
    错误文案不得泄露口令规则、端口、堆栈或内部路径。

    限流：受限端点在响应中返回 `X-RateLimit-*`，超限返回 429 与 `Retry-After`。
  contact:
    name: TeamLaunch 架构
  license:
    name: 内部项目
servers:
  - url: http://{host}:{port}/api/v1
    description: 局域网同步端点（端口由 UDP 17891 信标广播，可能漂移）
    variables:
      host:
        default: 192.168.1.10
      port:
        default: "17890"
security:
  - bearerAndSignature: []
  - {}
...
paths:
...
  /assets/{hash}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: 获取不可变资源（favicon / 管理员上传图标）
      description: hash 寻址，内容不可变，可安全永久缓存。
      security: [{}]
      parameters:
        - name: hash
          in: path
          required: true
          description: 资源内容的 SHA-256
          schema: { type: string, pattern: '^[a-f0-9]{64}$' }
      responses:
        '200':
          description: 资源二进制
          headers:
            Cache-Control:
              description: 固定为 public, max-age=31536000, immutable
              schema: { type: string }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            image/png: { schema: { type: string, format: binary } }
            image/x-icon: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Internal' }

  # ── 5 ────────────────────────────────────────────────────────────────────
  /auth/challenge:
    post:
      operationId: createChallenge
      tags: [auth]
      summary: 申请鉴权挑战
      security: [{}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChallengeRequest' }
      responses:
        '200':
          description: 返回挑战、盐值与 KDF 参数
          headers:
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ChallengeInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/RoleMismatch' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }

  # ── 6 ────────────────────────────────────────────────────────────────────
  /auth/verify:
    post:
      operationId: verifyChallenge
      tags: [auth]
      summary: 用 proof 换取短时令牌
      description: |
        口令明文永不上网：
        `proof = HMAC-SHA256(key=scrypt(password, salt, 32),
          msg = challengeId + ":" + challenge + ":" + deviceId + ":" + serverId)`。
        5 分钟内失败超过 10 次，该设备锁定 15 分钟。
      security: [{}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyRequest' }
      responses:
        '200':
          description: 认证通过
          headers:
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/TokenInfo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }

  # ── 8 ────────────────────────────────────────────────────────────────────
  /revisions:
    get:
      operationId: listRevisions
      tags: [publish]
      summary: 发布历史（含操作设备）
      security: [{ bearerAndSignature: [] }]
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 发布历史，从新到旧
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RevisionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }

  # ── 9 ────────────────────────────────────────────────────────────────────
  /revisions/{revision}/restore:
    post:
      operationId: restoreRevision
      tags: [publish]
      summary: 以指定历史版本的内容生成新的更高版本
      description: |
        不做版本回退：回滚本身也产生一个新的、更大的 revision，
        保证版本序列单调递增，客户端无需任何回退分支。
      security: [{ bearerAndSignature: [] }]
      parameters:
        - name: revision
          in: path
          required: true
          description: 要还原到的历史 revision
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestoreRequest' }
            example:
              baseRevision: 131
      responses:
        '200':
          description: 已生成新版本
          headers:
            X-TL-Revision: { $ref: '#/components/headers/Revision' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PublishResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 版本冲突或实例不匹配
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ConflictDetail' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }

  # ── 10 ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAndSignature:
      type: http
      scheme: bearer
      description: |
        Bearer 令牌 + 请求签名（ADR-006）。写请求必须同时携带：
          Authorization: Bearer <token>
          X-TL-Timestamp: <unix ms，服务端容许 ±120s>
          X-TL-Nonce:     <随机串，服务端 5 分钟内去重>
          X-TL-Signature: base64(HMAC-SHA256(
            key = scrypt(password, salt, 32),
            msg = token + "\n" + method + "\n" + path + "\n" + timestamp + "\n" + nonce + "\n" + sha256(rawBody)))
        缺少任一头部或校验不过即 401 ERR_SIGNATURE_INVALID / ERR_TOKEN_INVALID。
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: 单页条数上限
      schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
...
  headers:
    ETag:
      description: 强校验器 `<revision>-<contentHash前16位>`
      schema: { type: string }
    Revision:
      description: 当前配置版本号
      schema: { type: integer, format: int64 }
...
    RateLimitLimit:
      description: 当前窗口允许的请求数
      schema: { type: integer }
    RateLimitRemaining:
      description: 当前窗口剩余请求数
      schema: { type: integer }
    RateLimitReset:
      description: 窗口重置的 Unix 秒时间戳
      schema: { type: integer }
    RetryAfter:
      description: 建议重试等待秒数（429 时返回）
      schema: { type: integer }
  schemas:
    Envelope:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          description: 0 表示成功；非 0 见 error 字段的 ERR_* 机器码
          example: 0
        data:
          description: 业务数据；失败时为 null
          nullable: true
        message:
          type: string
          description: 人类可读描述。错误文案不得泄露口令规则、端口、堆栈或内部路径
        error:
          $ref: '#/components/schemas/ErrorCode'
        details:
          type: object
          description: 422 时附字段路径；其他错误可省略
          additionalProperties: true
        diagnosticId:
          type: string
          description: 错误诊断 ID，用于复制诊断与日志定位
    ErrorCode:
      type: string
      enum:
        - ERR_BAD_REQUEST
        - ERR_AUTH_REQUIRED
        - ERR_BAD_PROOF
        - ERR_CHALLENGE_EXPIRED
        - ERR_TOKEN_INVALID
        - ERR_SIGNATURE_INVALID
        - ERR_ROLE_MISMATCH
        - ERR_REVISION_CONFLICT
        - ERR_INSTANCE_MISMATCH
        - ERR_PAYLOAD_TOO_LARGE
        - ERR_VALIDATION_FAILED
        - ERR_RATE_LIMITED
        - ERR_NOT_FOUND
        - ERR_INTERNAL
...
    PublishResult:
      type: object
      required: [revision, contentHash, publishedAt]
      properties:
        revision: { type: integer, format: int64 }
        contentHash: { type: string }
        publishedAt: { type: string, format: date-time }
    RestoreRequest:
      type: object
      required: [baseRevision]
      description: |
        还原同样走乐观并发：baseRevision 必须是客户端打开还原对话框时看到的版本。
        否则 409，禁止静默覆盖他人在此期间发布的内容。
      properties:
        baseRevision: { type: integer, format: int64, description: 客户端已知的当前版本，用于乐观并发 }
    ConflictDetail:
      type: object
      properties:
        currentRevision: { type: integer, format: int64 }
        currentContentHash: { type: string }
    ChangeRecord:
      type: object
      required: [revision, publishedAt, summary, diff]
      properties:
        revision: { type: integer, format: int64 }
        publishedAt: { type: string, format: date-time }
        operatorDeviceId:
          type: [string, 'null']
          description: 匿名设备标识，非硬件信息
        summary: { type: string, maxLength: 200 }
        diff:
          type: object
          properties:
            added: { type: integer }
            updated: { type: integer }
            removed: { type: integer }
...
    RevisionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ChangeRecord' }
    ChallengeRequest:
      type: object
      required: [deviceId]
      properties:
        deviceId: { type: string, maxLength: 64 }
    ChallengeInfo:
      type: object
      required: [challengeId, challenge, salt, kdf, expiresAt, serverId]
      properties:
        challengeId: { type: string }
        challenge: { type: string, format: byte, description: 32 字节随机，base64 }
        salt: { type: string, format: byte }
        kdf:
          type: object
          properties:
            name: { type: string, enum: [scrypt] }
            N: { type: integer }
            r: { type: integer }
            p: { type: integer }
            keylen: { type: integer }
        expiresAt: { type: string, format: date-time, description: 60 秒后过期 }
        serverId: { type: string, format: uuid }
    VerifyRequest:
      type: object
      required: [challengeId, deviceId, proof]
      properties:
        challengeId: { type: string }
        deviceId: { type: string }
        proof:
          type: string
          format: byte
          description: HMAC-SHA256(verifier, challengeId:challenge:deviceId:serverId)
    TokenInfo:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time, description: 8 小时后过期，仅内存不落盘 }
...
  responses:
    BadRequest:
      description: 请求结构非法（ERR_BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Unauthorized:
      description: |
        鉴权失败：ERR_AUTH_REQUIRED / ERR_TOKEN_INVALID / ERR_SIGNATURE_INVALID /
        ERR_CHALLENGE_EXPIRED / ERR_BAD_PROOF
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    RoleMismatch:
      description: 目标不是管理员实例（ERR_ROLE_MISMATCH）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    NotFound:
      description: 资源不存在（ERR_NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
    RateLimited:
      description: 触发限流或设备被锁定（ERR_RATE_LIMITED）
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Internal:
      description: 未预期错误（ERR_INTERNAL），附 diagnosticId 供复制诊断
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }